backend/requirements.txt
fastapi
uvicorn[standard]
requests
starlette
redis
shapely
colorlog
pydantic
google-api-python-client
google-cloud-pubsub
google-cloud-firestore